Going to mirror Export like Import.
I'm going to have to create an ExportHandler class.

In the meantime, I added the parameters to Export to make it look like Import and added another queue, a bufferedExportQueue that will feed from the ExprotHanlder to the Export class.

However, it is currently a Queue<String> rather than a Queue of VarexpVariable. Mainly because you haven't created the varexpVariable object yet in Export.